requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
argcomplete==3.6.2
boto3==1.38.23
botocore==1.38.23
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
fasta2a==0.2.9
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
google-genai==1.16.1
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.25.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.61
langchain-google-genai==2.1.4
langchain-mcp-adapters==0.1.1
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
logfire-api==3.16.1
markdown-it-py==3.0.0
mcp==1.9.1
mdurl==0.1.2
mistralai==1.7.1
openai==1.82.0
opentelemetry-api==1.33.1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-ai==0.2.9
pydantic-ai-slim==0.2.9
pydantic-evals==0.2.9
pydantic-graph==0.2.9
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
zipp==3.22.0
zstandard==0.23.0
